streamlit_app/requirements.txt
streamlit>=1.28.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
tidb-vector>=0.0.9
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
